requirements.txt
ollama>=0.1.6
redis>=5.0.1
numpy>=1.24.0
PyMuPDF>=1.23.0  # for fitz
sentence-transformers>=2.2.2
streamlit>=1.32.0
psutil>=5.9.0
chromadb>=0.4.22
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
torch
transformers
Pillow
requests
